Rename WorkflowStatesGrid - #4712
